EEN Roles/requirements.txt
Flask==2.0.3
python-dotenv==0.19.2
requests==2.26.0
Werkzeug==2.0.3
